+++ modules/AWSPowerShell/Cmdlets/Mgn/AWS.Tools.Mgn.Completers.psm1
@@ -80,6 +80,13 @@ $MGN_Completers = {
 
     switch ($("$commandName/$parameterName"))
     {
+        # Amazon.Mgn.BootMode
+        "Update-MGNLaunchConfiguration/BootMode"
+        {
+            $v = "LEGACY_BIOS","UEFI"
+            break
+        }
+
         # Amazon.Mgn.ChangeServerLifeCycleStateSourceServerLifecycleState
         "Set-MGNServerLifeCycleState/LifeCycle_State"
         {
@@ -112,7 +119,7 @@ $MGN_Completers = {
             ($_ -eq "Update-MGNReplicationConfigurationTemplate/DefaultLargeStagingDiskType")
         }
         {
-            $v = "GP2","ST1"
+            $v = "GP2","GP3","ST1"
             break
         }
 
@@ -150,6 +157,7 @@ $MGN_Completers = {
 }
 
 $MGN_map = @{
+    "BootMode"=@("Update-MGNLaunchConfiguration")
     "DataPlaneRouting"=@("New-MGNReplicationConfigurationTemplate","Update-MGNReplicationConfiguration","Update-MGNReplicationConfigurationTemplate")
     "DefaultLargeStagingDiskType"=@("New-MGNReplicationConfigurationTemplate","Update-MGNReplicationConfiguration","Update-MGNReplicationConfigurationTemplate")
     "EbsEncryption"=@("New-MGNReplicationConfigurationTemplate","Update-MGNReplicationConfiguration","Update-MGNReplicationConfigurationTemplate")