+++ modules/AWSPowerShell/Cmdlets/Mgn/Basic/Complete-MGNCutover-Cmdlet.cs
@@ -33,10 +33,10 @@ namespace Amazon.PowerShell.Cmdlets.MGN
     /// will be terminated / deleted within 90 minutes. Launched Test or Cutover instances
     /// will NOT be terminated. The AWS Replication Agent will receive a command to uninstall
     /// itself (within 10 minutes). The following properties of the SourceServer will be changed
-    /// immediately: dataReplicationInfo.dataReplicationState will be to DISCONNECTED; The
-    /// SourceServer.lifeCycle.state will be changed to CUTOVER; The totalStorageBytes property
-    /// fo each of dataReplicationInfo.replicatedDisks will be set to zero; dataReplicationInfo.lagDuration
-    /// and dataReplicationInfo.lagDurationwill be nullified.
+    /// immediately: dataReplicationInfo.dataReplicationState will be changed to DISCONNECTED;
+    /// The SourceServer.lifeCycle.state will be changed to CUTOVER; The totalStorageBytes
+    /// property fo each of dataReplicationInfo.replicatedDisks will be set to zero; dataReplicationInfo.lagDuration
+    /// and dataReplicationInfo.lagDuration will be nullified.
     /// </summary>
     [Cmdlet("Complete", "MGNCutover", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.Mgn.Model.FinalizeCutoverResponse")]
@@ -50,7 +50,7 @@ namespace Amazon.PowerShell.Cmdlets.MGN
         #region Parameter SourceServerID
         /// <summary>
         /// <para>
-        /// <para>Request to finalize Cutover by Soure Server ID.</para>
+        /// <para>Request to finalize Cutover by Source Server ID.</para>
         /// </para>
         /// </summary>
         #if !MODULAR