+++ modules/AWSPowerShell/Cmdlets/RDS/Basic/Get-RDSDBCluster-Cmdlet.cs
@@ -37,9 +37,7 @@ namespace Amazon.PowerShell.Cmdlets.RDS
     /// What is Amazon Aurora?</a> in the <i>Amazon Aurora User Guide.</i></para><para>
     /// For more information on Multi-AZ DB clusters, see <a href="https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/multi-az-db-clusters-concepts.html">
     /// Multi-AZ deployments with two readable standby DB instances</a> in the <i>Amazon RDS
-    /// User Guide.</i></para><note><para>
-    /// The Multi-AZ DB clusters feature is in preview and is subject to change.
-    /// </para></note><para>
+    /// User Guide.</i></para><para>
     /// This operation can also return information for Amazon Neptune DB instances and Amazon
     /// DocumentDB instances.
     /// </para><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.