+++ modules/AWSPowerShell/Cmdlets/RDS/Basic/Remove-RDSDBCluster-Cmdlet.cs
@@ -38,9 +38,7 @@ namespace Amazon.PowerShell.Cmdlets.RDS
     /// What is Amazon Aurora?</a> in the <i>Amazon Aurora User Guide.</i></para><para>
     /// For more information on Multi-AZ DB clusters, see <a href="https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/multi-az-db-clusters-concepts.html">
     /// Multi-AZ deployments with two readable standby DB instances</a> in the <i>Amazon RDS
-    /// User Guide.</i></para><note><para>
-    /// The Multi-AZ DB clusters feature is in preview and is subject to change.
-    /// </para></note>
+    /// User Guide.</i></para>
     /// </summary>
     [Cmdlet("Remove", "RDSDBCluster", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.High)]
     [OutputType("Amazon.RDS.Model.DBCluster")]